requirements.txt
openai
gtts
moviepy
diffusers
transformers
accelerate
torch
google-generativeai
